feat(schema): enhance rating precision and extend show metadata
- Change UserRating type to double for better precision in UserShows, UserMovies, and WatchedEpisodes.
- Add VoteAverage, ProductionStatus, LatestEpisodeInfo, and NextAirDate columns to enrich show details.
- Implement new method in ITmdbService to fetch TV season details.
- Update DbContextModelSnapshot to reflect schema changes.

+++ Vivaply.API/Entities/Entertainment/WatchedEpisode.cs
@@ -17,7 +17,7 @@ namespace Vivaply.API.Entities.Entertainment
         public DateTime WatchedAt { get; set; } = DateTime.UtcNow;
 
         [Range(1, 10)]
-        public int? UserRating { get; set; }
+        public double? UserRating { get; set; }
 
         [MaxLength(500)]
         public string? Review { get; set; }
